refactor(bench): separate component app painter
